% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick_conf_mat_plot.R
\name{ggconfusion_matrix}
\alias{ggconfusion_matrix}
\title{Confusion Matrix Heatmap Plot for Categorical Data}
\usage{
ggconfusion_matrix(
  object,
  colors = c("#00BFC4", "#7CAE00", "#F8755D"),
  ggtheme = ggplot2::theme_minimal,
  outline.color = "gray20"
)
}
\arguments{
\item{object}{The `conf_mat` data frame returned from `conf_mat()`.}

\item{colors}{a vector of 3 colors for the most correct, mid-point, and most miss-classified frequencies of predictions in the heatmap.}

\item{ggtheme}{ggplot2 function or theme object. Default value is
`theme_minimal`. Other values include the official ggplot2 themes
theme_gray, theme_bw, theme_minimal, theme_classic, theme_void, .... Theme
objects are also allowed (e.g., `theme_classic()`).}

\item{outline.color}{the outline color of square or circle. Default value is
"gray20".}
}
\value{
A ggplot2 plot object
}
\description{
Calculates a cross-tabulation graphic of observed and predicted classes. A [ggplot2::geom_tile()] graphic.
}
