% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{my_ggplot_option_settings}
\alias{my_ggplot_option_settings}
\title{my_ggplot_option_settings() for loading in .Rprofile}
\usage{
my_ggplot_option_settings(...)
}
\arguments{
\item{...}{Other arguments passed to \code{theme_minimal}}
}
\description{
my_ggplot_option_settings() for loading in .Rprofile
}
