% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withfreq.R
\name{withfreq}
\alias{withfreq}
\title{a categorical count helper function for cleaner ggplots}
\usage{
withfreq(string, width = 20)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something
coercible to one.}

\item{width}{character vector wrap for better ggplot axis labeling

 @examples

 x <- factor(letters[rpois(100, 5)])

 withfreq(x) |> table()}
}
\description{
a categorical count helper function for cleaner ggplots
}
